requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
gradio
